# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Flight search and preliminary booking API with unified success and error envelopes.",
        "title": "FlySoft Flight Integration Service API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/v1/bookings": {
            "post": {
                "description": "Creates a preliminary booking for an existing, unexpired offer.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookings"
                ],
                "summary": "Create a preliminary booking",
                "parameters": [
                    {
                        "description": "Booking request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.BookingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.BookingEnvelope"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorEnvelope"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/api/v1/flights/search": {
            "post": {
                "description": "Searches the mock flight provider, prices matching offers, persists them, and returns normalized offers.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "flights"
                ],
                "summary": "Search flight offers",
                "parameters": [
                    {
                        "description": "Flight search request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SearchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SearchEnvelope"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorEnvelope"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Returns service liveness status.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.HealthResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorEnvelope"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.BookingEnvelope": {
            "type": "object",
            "properties": {
                "booking": {
                    "$ref": "#/definitions/dto.BookingResponse"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "dto.BookingRequest": {
            "type": "object",
            "properties": {
                "offer_id": {
                    "type": "string",
                    "example": "OF-123456"
                },
                "passengers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.PassengerDTO"
                    }
                }
            }
        },
        "dto.BookingResponse": {
            "type": "object",
            "properties": {
                "booking_id": {
                    "type": "string",
                    "example": "BK-987654"
                },
                "expires_at": {
                    "type": "string",
                    "example": "2026-05-28T15:30:00Z"
                },
                "offer_id": {
                    "type": "string",
                    "example": "OF-123456"
                },
                "status": {
                    "type": "string",
                    "example": "created"
                }
            }
        },
        "dto.ErrorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "VALIDATION_ERROR"
                },
                "message": {
                    "type": "string",
                    "example": "from is required"
                }
            }
        },
        "dto.ErrorEnvelope": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/dto.ErrorBody"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "dto.HealthResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "dto.OfferResponse": {
            "type": "object",
            "properties": {
                "airline": {
                    "type": "string",
                    "example": "TK"
                },
                "base_price": {
                    "type": "number",
                    "example": 500
                },
                "commission": {
                    "type": "number",
                    "example": 25
                },
                "currency": {
                    "type": "string",
                    "example": "USD"
                },
                "departure_date": {
                    "type": "string",
                    "example": "2026-06-15"
                },
                "flight_number": {
                    "type": "string",
                    "example": "TK255"
                },
                "from": {
                    "type": "string",
                    "example": "DYU"
                },
                "offer_id": {
                    "type": "string",
                    "example": "OF-123456"
                },
                "profit": {
                    "type": "number",
                    "example": 50
                },
                "provider": {
                    "type": "string",
                    "example": "MockAvia"
                },
                "return_date": {
                    "type": "string",
                    "example": "2026-06-25"
                },
                "service_fee": {
                    "type": "number",
                    "example": 25
                },
                "to": {
                    "type": "string",
                    "example": "IST"
                },
                "total_price": {
                    "type": "number",
                    "example": 550
                }
            }
        },
        "dto.PassengerDTO": {
            "type": "object",
            "properties": {
                "document_number": {
                    "type": "string",
                    "example": "A1234567"
                },
                "first_name": {
                    "type": "string",
                    "example": "Alisher"
                },
                "last_name": {
                    "type": "string",
                    "example": "Sabirov"
                },
                "type": {
                    "type": "string",
                    "example": "adult"
                }
            }
        },
        "dto.SearchEnvelope": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.OfferResponse"
                    }
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "dto.SearchRequest": {
            "type": "object",
            "properties": {
                "adults": {
                    "type": "integer",
                    "example": 1
                },
                "children": {
                    "type": "integer",
                    "example": 1
                },
                "currency": {
                    "type": "string",
                    "example": "USD"
                },
                "departure_date": {
                    "type": "string",
                    "example": "2026-06-15"
                },
                "from": {
                    "type": "string",
                    "example": "DYU"
                },
                "infants": {
                    "type": "integer",
                    "example": 0
                },
                "return_date": {
                    "type": "string",
                    "example": "2026-06-25"
                },
                "to": {
                    "type": "string",
                    "example": "IST"
                }
            }
        }
    }
}